session modification for payments methods

+++ src/xbridge/util/xutil.h
@@ -68,11 +68,11 @@ using namespace json_spirit;
     boost::uint64_t timeToInt(const bpt::ptime &time);
     bpt::ptime intToTime(const uint64_t& number);
 
-    double xBridgeValueFromAmount(uint64_t amount);
+    double xBridgeValueFromAmount(uint256 amount);
 
     uint64_t xBridgeAmountFromReal(double val);
     std::string xBridgeStringValueFromPrice(double price);
-    std::string xBridgeStringValueFromAmount(uint64_t amount);
+    std::string xBridgeStringValueFromAmount(uint256 amount);
 
     /**
      * @brief Returns true if the input precision is supported by xbridge.
